+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # dmptool-api Changelog
 
+## 2026-05-14
+- Added helper types to `src/types.ts` to facilitate access to nested portions of the JSON schemas
+- Updated the `server.ts` to load the new GraphQL plugin
+- Modified the api v3 routes to allow creation of DMPs
+- Update the serialization plugin to support the DMP Tool and RDA schemas
+- Added a redirect to the routes plugin to handle missing slash to Swagger UI
+- Fixed some issues with the structure of the DMP Tool schema
+- Added the refresh token to the Fastify request
+- Added Plan, Project and VersionedTemplate models
+- Added an abstract `BaseGQL.ts` file to handle common GraphQL functionality
+- Added GraphQL query and mutation documents to `src/graphql/`
+- Added a new GraphQL plugin to set up the GraphQL client
+- Updated the `configuration.ts` to support GraphQL config
+- Added JSON sample files to `docs/jsonSamples/` for use when testing the Swagger UI
+- Added a `generate` script to package.json to generate the GraphQL schema
+- Added a GraphQL `codegen.ts` file to generate the GraphQL schema to `src/generated/`
+- Updated `.gitignore` to skip the `src/generated` directory
+- Added Graphql URI to the `docker-compose.yml`
+- Added `generateDMPId` function to `maDMP.ts` to support Current DMP Tool integration
+- Updated dependencies
+
 ## 2026-04-02
 - Updated the `server` to use the new plugins and added `strictSchema: false` to allow our Zod defaults to work
 - Added `@fastify/rate-limit`, `@fastify/swagger` and `@fastify/swagger-ui`